Change: Move update_from_applied_settings to branch API
The update_from_applied_settings function might be useful for all
callers of the update_branch_protection API method if only some settings
should be changed and not all.
